refactor(core): consolidate database schemas into unified structure
- Rename schema files to schema.sql.ts for clarity and consistency
- Create central database/schema.sql.ts file that exports all domain schemas
- Update import paths in article and site domain modules
- Improve drizzle-kit configuration and schema management
