+++ packages/core/src/database/schema.sql.ts
@@ -1,0 +1,5 @@
+// Combined schema file for drizzle-kit
+// This file imports all domain-specific schemas
+
+export * from "../article/schema.sql";
+export * from "../site/schema.sql";
